Created "events" folder
